Refactor appendPathComponent
